requirements.txt
streamlit>=1.33
openai>=1.30
faiss-cpu>=1.8
pypdf>=4.2
numpy>=1.26
python-dotenv>=1.0
